crawler/requirements.txt
ccxt==4.1.75
psycopg2-binary==2.9.9
pandas==2.1.1

# Use for time zone conversion
pytz==2023.3